EjemploPrueba/requirements.txt
faust==1.10.4
robinhood-aiokafka==1.1.6
pytz
shapely